resolved conversations, formatting, unnecessary code, etc.
